Create Top_view.cpp
Problem Description

Given a binary tree of integers denoted by root A. Return an array of integers representing the top view of the Binary tree.

Top view of a Binary Tree is a set of nodes visible when the tree is visited from top.

Return the nodes in any order.



Problem Constraints

1 <= Number of nodes in binary tree <= 100000

0 <= node values <= 10^9



Input Format

First and only argument is head of the binary tree A.



Output Format

Return an array, representing the top view of the binary tree.



Example Input

Input 1:

 
            1
          /   \
         2    3
        / \  / \
       4   5 6  7
      /
     8 

Input 2:

 
            1
           /  \
          2    3
           \
            4
             \
              5



Example Output

Output 1:

 [1, 2, 4, 8, 3, 7]

Output 2:

 [1, 2, 3]



Example Explanation

Explanation 1:

Top view is described.

Explanation 2:

Top view is described.


Solution



Print top view of a binary tree

Given a binary tree, print the top view of it. Assume the left and right child of a node makes a 45–degree angle with the parent.

For example, the top view of the following tree is 2, 1, 3, 6:

 
Top View of Binary Tree

    Practice this problem

We can easily solve this problem with the help of hashing. The idea is to create an empty map where each key represents the relative horizontal distance of the node from the root node, and the value in the map maintains a pair containing the node’s value and its level number. Then perform preorder traversal on the tree. Suppose the current node’s level is less than the maximum level seen so far for the same horizontal distance as the current node or current horizontal distance is seen for the first time. In that case, update the value and the level for the current horizontal distance in the map. For each node, recur for its left subtree by decreasing horizontal distance and increasing level by one, and recur for right subtree by increasing both level and horizontal distance by one.

 
The following figure shows the horizontal distance and level of each node in the above binary tree:

Horizontal distance vs Level – Binary Tree

The final values in the map will be:
horizontal distance —> (node’s value, node’s level)
-1 —> (2, 2)
0 —> (1, 1)
1 —> (3, 2)
2 —> (6, 3)

The algorithm can be implemented as follows in C++, Java, and Python: